+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Info: improved state/place display (requires rescan, limited to AU/GB/IN/US)
 - Info: edit tags with state placeholder
 - Countries: show states for selected countries
+- Tags: delete selected tags from all media in collection
 - improved support for system font scale
 
 ### Changed